split SlackRecord into SlackFormatter
moves the message formatting parts of SlackRecord into SlackFormatter
this allows the SlackFormatter to be swapped out by users to create new message formats
