+++ src/Monolog/Handler/Slack/SlackRecord.php
@@ -11,10 +11,8 @@
 
 namespace Monolog\Handler\Slack;
 
-use Monolog\Logger;
-use Monolog\Utils;
-use Monolog\Formatter\NormalizerFormatter;
-use Monolog\Formatter\FormatterInterface;
+use InvalidArgumentException;
+use Monolog\Formatter\SlackFormatter;
 
 /**
  * Slack record utility helping to log to Slack webhooks or API.
@@ -26,12 +24,16 @@ use Monolog\Formatter\FormatterInterface;
  */
 class SlackRecord
 {
+    /** @deprecated Use {@see SlackFormatter::COLOR_DANGER} */
     public const COLOR_DANGER = 'danger';
 
+    /** @deprecated Use {@see SlackFormatter::COLOR_WARNING} */
     public const COLOR_WARNING = 'warning';
 
+    /** @deprecated Use {@see SlackFormatter::COLOR_GOOD} */
     public const COLOR_GOOD = 'good';
 
+    /** @deprecated Use {@see SlackFormatter::COLOR_DEFAULT} */
     public const COLOR_DEFAULT = '#e3e4e6';
 
     /**
@@ -52,63 +54,15 @@ class SlackRecord
      */
     private $userIcon;
 
-    /**
-     * Whether the message should be added to Slack as attachment (plain text otherwise)
-     * @var bool
-     */
-    private $useAttachment;
-
-    /**
-     * Whether the the context/extra messages added to Slack as attachments are in a short style
-     * @var bool
-     */
-    private $useShortAttachment;
-
-    /**
-     * Whether the attachment should include context and extra data
-     * @var bool
-     */
-    private $includeContextAndExtra;
-
-    /**
-     * Dot separated list of fields to exclude from slack message. E.g. ['context.field1', 'extra.field2']
-     * @var array
-     */
-    private $excludeFields;
-
-    /**
-     * @var FormatterInterface
-     */
-    private $formatter;
-
-    /**
-     * @var NormalizerFormatter
-     */
-    private $normalizerFormatter;
-
     public function __construct(
         ?string $channel = null,
         ?string $username = null,
-        bool $useAttachment = true,
-        ?string $userIcon = null,
-        bool $useShortAttachment = false,
-        bool $includeContextAndExtra = false,
-        array $excludeFields = array(),
-        FormatterInterface $formatter = null
+        ?string $userIcon = null
     ) {
         $this
             ->setChannel($channel)
             ->setUsername($username)
-            ->useAttachment($useAttachment)
-            ->setUserIcon($userIcon)
-            ->useShortAttachment($useShortAttachment)
-            ->includeContextAndExtra($includeContextAndExtra)
-            ->excludeFields($excludeFields)
-            ->setFormatter($formatter);
-
-        if ($this->includeContextAndExtra) {
-            $this->normalizerFormatter = new NormalizerFormatter();
-        }
+            ->setUserIcon($userIcon);
     }
 
     /**
@@ -118,7 +72,6 @@ class SlackRecord
     public function getSlackData(array $record): array
     {
         $dataArray = array();
-        $record = $this->removeExcludedFields($record);
 
         if ($this->username) {
             $dataArray['username'] = $this->username;
@@ -128,53 +81,18 @@ class SlackRecord
             $dataArray['channel'] = $this->channel;
         }
 
-        if ($this->formatter && !$this->useAttachment) {
-            $message = $this->formatter->format($record);
-        } else {
-            $message = $record['message'];
-        }
+        // Add the slack-formatted message to the json body
+        $message = $record['formatted'] ?? $record['message'];
 
-        if ($this->useAttachment) {
-            $attachment = array(
-                'fallback'  => $message,
-                'text'      => $message,
-                'color'     => $this->getAttachmentColor($record['level']),
-                'fields'    => array(),
-                'mrkdwn_in' => array('fields'),
-                'ts'        => $record['datetime']->getTimestamp(),
+        if (is_array($message)) {
+            $dataArray = array_merge($dataArray, $message);
+        } else if (is_scalar($message)) {
+            $dataArray['text'] = (string) $message;
+        } else {
+            throw new InvalidArgumentException(
+                'Expected formatter to return a scalar or a slack message array. Instead got type '
+                . gettype($message)
             );
-
-            if ($this->useShortAttachment) {
-                $attachment['title'] = $record['level_name'];
-            } else {
-                $attachment['title'] = 'Message';
-                $attachment['fields'][] = $this->generateAttachmentField('Level', $record['level_name']);
-            }
-
-            if ($this->includeContextAndExtra) {
-                foreach (array('extra', 'context') as $key) {
-                    if (empty($record[$key])) {
-                        continue;
-                    }
-
-                    if ($this->useShortAttachment) {
-                        $attachment['fields'][] = $this->generateAttachmentField(
-                            (string) $key,
-                            $record[$key]
-                        );
-                    } else {
-                        // Add all extra fields as individual fields in attachment
-                        $attachment['fields'] = array_merge(
-                            $attachment['fields'],
-                            $this->generateAttachmentFields($record[$key])
-                        );
-                    }
-                }
-            }
-
-            $dataArray['attachments'] = array($attachment);
-        } else {
-            $dataArray['text'] = $message;
         }
 
         if ($this->userIcon) {
@@ -189,44 +107,7 @@ class SlackRecord
     }
 
     /**
-     * Returns a Slack message attachment color associated with
-     * provided level.
-     */
-    public function getAttachmentColor(int $level): string
-    {
-        switch (true) {
-            case $level >= Logger::ERROR:
-                return static::COLOR_DANGER;
-            case $level >= Logger::WARNING:
-                return static::COLOR_WARNING;
-            case $level >= Logger::INFO:
-                return static::COLOR_GOOD;
-            default:
-                return static::COLOR_DEFAULT;
-        }
-    }
-
-    /**
-     * Stringifies an array of key/value pairs to be used in attachment fields
-     */
-    public function stringify(array $fields): string
-    {
-        $normalized = $this->normalizerFormatter->format($fields);
-
-        $hasSecondDimension = count(array_filter($normalized, 'is_array'));
-        $hasNonNumericKeys = !count(array_filter(array_keys($normalized), 'is_numeric'));
-
-        return $hasSecondDimension || $hasNonNumericKeys
-            ? Utils::jsonEncode($normalized, JSON_PRETTY_PRINT|Utils::DEFAULT_JSON_FLAGS)
-            : Utils::jsonEncode($normalized, Utils::DEFAULT_JSON_FLAGS);
-    }
-
-    /**
      * Channel used by the bot when posting
-     *
-     * @param ?string $channel
-     *
-     * @return SlackHandler
      */
     public function setChannel(?string $channel = null): self
     {
@@ -237,21 +118,10 @@ class SlackRecord
 
     /**
      * Username used by the bot when posting
-     *
-     * @param ?string $username
-     *
-     * @return SlackHandler
      */
     public function setUsername(?string $username = null): self
     {
         $this->username = $username;
-
-        return $this;
-    }
-
-    public function useAttachment(bool $useAttachment = true): self
-    {
-        $this->useAttachment = $useAttachment;
 
         return $this;
     }
@@ -265,92 +135,5 @@ class SlackRecord
         }
 
         return $this;
-    }
-
-    public function useShortAttachment(bool $useShortAttachment = false): self
-    {
-        $this->useShortAttachment = $useShortAttachment;
-
-        return $this;
-    }
-
-    public function includeContextAndExtra(bool $includeContextAndExtra = false): self
-    {
-        $this->includeContextAndExtra = $includeContextAndExtra;
-
-        if ($this->includeContextAndExtra) {
-            $this->normalizerFormatter = new NormalizerFormatter();
-        }
-
-        return $this;
-    }
-
-    public function excludeFields(array $excludeFields = []): self
-    {
-        $this->excludeFields = $excludeFields;
-
-        return $this;
-    }
-
-    public function setFormatter(?FormatterInterface $formatter = null): self
-    {
-        $this->formatter = $formatter;
-
-        return $this;
-    }
-
-    /**
-     * Generates attachment field
-     *
-     * @param string|array $value
-     */
-    private function generateAttachmentField(string $title, $value): array
-    {
-        $value = is_array($value)
-            ? sprintf('```%s```', substr($this->stringify($value), 0, 1990))
-            : $value;
-
-        return array(
-            'title' => ucfirst($title),
-            'value' => $value,
-            'short' => false,
-        );
-    }
-
-    /**
-     * Generates a collection of attachment fields from array
-     */
-    private function generateAttachmentFields(array $data): array
-    {
-        $fields = array();
-        foreach ($this->normalizerFormatter->format($data) as $key => $value) {
-            $fields[] = $this->generateAttachmentField((string) $key, $value);
-        }
-
-        return $fields;
-    }
-
-    /**
-     * Get a copy of record with fields excluded according to $this->excludeFields
-     */
-    private function removeExcludedFields(array $record): array
-    {
-        foreach ($this->excludeFields as $field) {
-            $keys = explode('.', $field);
-            $node = &$record;
-            $lastKey = end($keys);
-            foreach ($keys as $key) {
-                if (!isset($node[$key])) {
-                    break;
-                }
-                if ($lastKey === $key) {
-                    unset($node[$key]);
-                    break;
-                }
-                $node = &$node[$key];
-            }
-        }
-
-        return $record;
     }
 }
